Added new class to listen for wifi changes so a card can be displayed on watch. Moved all code relating to wear from utils into new class.
